+++ src/pages/Todo/List/TodoListPagination.js
@@ -4,8 +4,8 @@ import { usePaginatedGet } from '../../../hooks/useRequest';
 import { List } from '@ui5/webcomponents-react/lib/List';
 import { StandardListItem } from '@ui5/webcomponents-react/lib/StandardListItem';
 import { Spinner } from '@ui5/webcomponents-react';
+import { Pagination } from '../../../components/Pagination/Pagination';
 import Constants from '../../../util/Constants';
-import { Pagination } from '../../../components/ViewSetting/Pagination';
 
 export default function TodoListPagination() {
   const [page, setPage] = useState(0);
